Add files via upload
In Python, a class is a blueprint for creating objects. Objects are instances of a class, and a class defines a set of attributes and methods that characterize those objects. Classes are a fundamental concept in object-oriented programming (OOP), which is a programming paradigm that revolves around the concept of objects.

Here's a brief breakdown of the key components of a Python class:

Attributes:

Attributes are variables that store data. They represent the characteristics or properties of an object. In a class, attributes are defined within the class and are associated with instances (objects) of that class.

Methods:

Methods are functions that are defined within a class and operate on the attributes of instances of that class. They represent the behaviors or actions that objects of the class can perform.

Constructor (__init__ method):

The __init__ method is a special method in Python classes that is called when an object is created. It initializes the attributes of the object.

Self:

self is a convention in Python used as the first parameter of methods within a class. It refers to the instance of the class, allowing you to access and modify its attributes.

Instance:

An instance is an individual object created from a class. Each instance has its own set of attributes and can call the methods defined in the class.
python